refactor: move starredSpaces to useAccount
