+++ apps/ui/src/components/App/Sidebar.vue
@@ -2,7 +2,7 @@
 import draggable from 'vuedraggable';
 
 const uiStore = useUiStore();
-const spacesStore = useSpacesStore();
+const { starredSpacesLoaded, starredSpaces } = useAccount();
 </script>
 
 <template>
@@ -10,10 +10,10 @@ const spacesStore = useSpacesStore();
     <router-link :to="{ name: 'landing' }" class="h-[72px] block">
       <IH-stop class="inline-block my-4 w-[32px] h-[32px] text-skin-link" />
     </router-link>
-    <UiLoading v-if="!spacesStore.starredSpacesLoaded" />
+    <UiLoading v-if="!starredSpacesLoaded" />
     <draggable
       v-else
-      v-model="spacesStore.starredSpaces"
+      v-model="starredSpaces"
       :delay="100"
       :delay-on-touch-only="true"
       :touch-start-threshold="35"
